+++ lotti/lib/pages/my_day.dart
@@ -1,9 +1,9 @@
+import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:lotti/classes/journal_entities.dart';
 import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/theme.dart';
-import 'package:lotti/widgets/journal/entry_detail_route.dart';
 import 'package:lotti/widgets/journal/entry_tools.dart';
 import 'package:syncfusion_flutter_calendar/calendar.dart';
 
@@ -89,15 +89,8 @@ class MyDayPage extends StatelessWidget {
           onTap: (CalendarTapDetails cal) {
             cal.appointments?.forEach((element) {
               Meeting meeting = element as Meeting;
-              debugPrint(meeting.journalEntity.toString());
-
-              Navigator.of(context).push(
-                MaterialPageRoute(
-                  builder: (BuildContext context) {
-                    return EntryDetailRoute(item: meeting.journalEntity);
-                  },
-                ),
-              );
+              String entryId = meeting.journalEntity.meta.id;
+              context.router.pushNamed('/journal/$entryId');
             });
           },
         );
