+++ lotti/lib/routes/router.dart
@@ -1,11 +1,11 @@
 import 'package:auto_route/auto_route.dart';
-import 'package:lotti/pages/flagged_entries_page.dart';
 import 'package:lotti/pages/home_page.dart';
-import 'package:lotti/pages/journal_page.dart';
 import 'package:lotti/pages/my_day.dart';
-import 'package:lotti/pages/tasks_page.dart';
 import 'package:lotti/routes/dashboard_routes.dart';
+import 'package:lotti/routes/flagged_routes.dart';
+import 'package:lotti/routes/journal_routes.dart';
 import 'package:lotti/routes/settings_routes.dart';
+import 'package:lotti/routes/task_routes.dart';
 
 @MaterialAutoRouter(
   replaceInRouteName: 'Page,Route',
@@ -14,47 +14,9 @@ import 'package:lotti/routes/settings_routes.dart';
       path: '/',
       page: HomePage,
       children: [
-        AutoRoute(
-          path: 'journal',
-          name: 'JournalRouter',
-          page: EmptyRouterPage,
-          children: [
-            AutoRoute(
-              path: '',
-              page: JournalPage,
-            ),
-          ],
-        ),
-        AutoRoute(
-          path: 'flagged',
-          name: 'FlaggedRouter',
-          page: EmptyRouterPage,
-          children: [
-            AutoRoute(
-              path: '',
-              page: FlaggedEntriesPage,
-            ),
-            AutoRoute(
-              path: ':userId',
-              page: EmptyRouterPage,
-            ),
-          ],
-        ),
-        AutoRoute(
-          path: 'tasks',
-          name: 'TasksRouter',
-          page: EmptyRouterPage,
-          children: [
-            AutoRoute(
-              path: '',
-              page: TasksPage,
-            ),
-            AutoRoute(
-              path: ':userId',
-              page: EmptyRouterPage,
-            ),
-          ],
-        ),
+        journalRoutes,
+        flaggedRoutes,
+        taskRoutes,
         dashboardRoutes,
         AutoRoute(
           path: 'my_day',
